--- a/internal/cache/file/downloader/job.go
+++ b/internal/cache/file/downloader/job.go
@@ -382,7 +382,7 @@ func (job *Job) createCacheFile() (*os.File, error) {
 	var err error
 	// Try using O_DIRECT while opening file when parallel downloads are enabled
 	// and O_DIRECT use is not disabled.
-	if job.fileCacheConfig.EnableParallelDownloads && !job.fileCacheConfig.EnableODirect {
+	if job.fileCacheConfig.EnableParallelDownloads && job.fileCacheConfig.EnableODirect {
 		cacheFile, err = cacheutil.CreateFile(job.fileSpec, openFileFlags|syscall.O_DIRECT)
 		if errors.Is(err, fs.ErrInvalid) || errors.Is(err, syscall.EINVAL) {
 			logger.Warnf("downloadObjectAsync: failure in opening file with O_DIRECT, falling back to without O_DIRECT")