--- a/internal/cache/file/downloader/parallel_downloads_job.go
+++ b/internal/cache/file/downloader/parallel_downloads_job.go
@@ -62,10 +62,8 @@ func (job *Job) downloadRange(ctx context.Context, dstWriter io.Writer, start, e
 
 	monitor.CaptureGCSReadMetrics(ctx, util.Parallel, end-start)
 
-	// Use of memory aligned buffer is not required if use of O_DIRECT is disabled.
+	// Use memory aligned buffer if O_DIRECT is enabled.
 	if job.fileCacheConfig.EnableODirect {
-		_, err = io.CopyN(dstWriter, newReader, end-start)
-	} else {
 		_, err = cacheutil.CopyUsingMemoryAlignedBuffer(ctx, newReader, dstWriter, end-start,
 			job.fileCacheConfig.WriteBufferSize)
 		// If context is canceled while reading/writing in CopyUsingMemoryAlignedBuffer
@@ -74,6 +72,8 @@ func (job *Job) downloadRange(ctx context.Context, dstWriter io.Writer, start, e
 		if !errors.Is(err, context.Canceled) && errors.Is(ctx.Err(), context.Canceled) {
 			err = errors.Join(err, ctx.Err())
 		}
+	} else {
+		_, err = io.CopyN(dstWriter, newReader, end-start)
 	}
 
 	if err != nil {