### Description
Fix to not use O_DIRECT by default.

Tests performed manually:
machine spec:
n2-standard-96
16LSSDs

fio
```
; -- use nrfiles and rw to CLI args to control readtype and number of files --
[global]
ioengine=libaio
direct=0
fadvise_hint=0
iodepth=64
invalidate=1
nrfiles=16
thread=1
openfiles=1
group_reporting=1
create_serialize=0
allrandrepeat=0
file_service_type=random
numjobs=1
filename_format=$jobname.$jobnum/$filenum

[Workload]
directory=${DIR}
bs=128K
filesize=10G
rw=read
```

pytorch data loader
```
Loaded Llama 70B model
```

With `enable-o-direct=true` 
Fio throughput: 733MiB/s (221s)
Llama 70B model load time: 275s

Without `enable-o-direct=false` (default with this fix)
Fio throughput: 812MiB/s (201s)
Llama 70B model load time: 267s



### Link to the issue in case of a bug fix.
NA

### Testing details
1. Manual - NA
2. Unit tests - NA
3. Integration tests - NA